lib: cleanups post lnodes renames

+++ include/silofs/fs/inode.h
@@ -93,7 +93,7 @@ void silofs_ii_update_iblocks(struct silofs_inode_info *ii,
                               enum silofs_ltype ltype, long dif);
 
 void silofs_ii_update_isize(struct silofs_inode_info *ii,
-                            const struct silofs_creds *creds, loff_t size);
+                            const struct silofs_creds *creds, ssize_t size);
 
 void silofs_ii_update_iattrs(struct silofs_inode_info *ii,
                              const struct silofs_creds *creds,
